Glissa

Documentation for Glissa.

Glissa.CubicSpline

Glissa.DividedDifference

Glissa.NewtonForm

Glissa.NewtonForm

Base.:*

Glissa.cubic_spline_coefficients!

Glissa.deboor_swartz_criterion

Glissa.derivative

Glissa.divided_difference

Glissa.global_derivatives!

Glissa.horner_in_interval

Glissa.hyman_coefficients!

Glissa.hyman_criterion

Glissa.integral_in_interval

Glissa.integrate

Glissa.normalized_bspline

Glissa.project_to_monotonicity!

Glissa.single_polynomial

Glissa.CubicSpline — Type

This is a traditional cubic spline. It has N+1 points on the abcissa, $\tau$. It has N cubic interpolants in c, which is size 4 x N. At each interval, the polynomial is in Horner Form,

$P_i(x) = c_{1,i} + c_{2,i}(x-x_i) + c_{3,i}(x-x_i)^2 + c_{4,i}(x-x_i)^3$

This spline has one type for the abcissa and one for the ordinate coefficients. It should be the case that one(T) * one(X) is of type T.

This can be a spline of another order. It's just a matter of the dimensions of the coefficient array, c.

Glissa.DividedDifference — Type

A recursive implementation of divided differences.

Glissa.NewtonForm — Type

The Newton form for an interpolating polynomial represents $f(x)$ at points $\tau$` by a sum.

``\sum{i=0}^n f[x0,\ldots,xi]\prod{j=0}^{i-1}(x-x_j)

Here, the a are the divided differences and $\tau$ are the x[j].

Glissa.NewtonForm — Method

Construct a NewtonForm. From Conte and deBoor 1980, Eqn. 4.8.

Base.:* — Method

Multiply two splines of order m1 and m2 to get a spline of order m1+m2. This only works if the two splines have the same abcissa.

Glissa.cubic_spline_coefficients! — Method

Given an abcissa, τ, of length N+1 values f of length N+1, and derivatives s of length N+1, this computes the 4xN matrix c of cubic coefficients. Conte and deBoor Eq. 4.55

Glissa.deboor_swartz_criterion — Method

"Accurate Monotonicity Preserving Cubic Interpolation" by James M Hyman. 1983.

Glissa.derivative — Method

Given a spline, create a new spline that is its derivative.

Glissa.divided_difference — Method

Calculate divided differences of a function f at points τ. This calculates the divided difference from τi to τj. The dd is a memoization.

Glissa.global_derivatives! — Method

τ are the abcissa from 1 to N+1. f are the known values of length N+1. fp_bounds are derivatives at the first and last endpoint. fp is the output, of length N + 1. This is from Conte and deBoor Sec. 4.8, Eq. 4.60. The equation is this, with $i$ running from 2 to N.

$(\Delta x_i) s_{i-1}+2(\Delta x_{i-1}+\Delta x_i)s_i +(\Delta x_{i-1}s_{i+1}) = 3(f[x_{i-1},x_i]\Delta x_i+f[x_i,x_{i+1}]\Delta x_{i-1})$

Here, $\Delta x_i = x_{i+1} -x{i}$ and

$f[x_i,x_{i+1}] = \frac{f[x_{i+1}] - f[x_i]}{x_{i+1}-x_i}$

Glissa.horner_in_interval — Method

((cs.c[4, i] * Δ + cs.c[3, i]) * Δ + cs.c[2, i]) * Δ + cs.c[1, i]

Glissa.hyman_coefficients! — Method

This is how splinefun sets the coefficients. Seems to have the same result as the code above.

Glissa.hyman_criterion — Method

This comes from Hyman's paper but is modified by the R implementation of splinefun, which Simon Wood wrote. I didn't understand from the paper the exact structure of decisions to take in the function.

Glissa.integral_in_interval — Method

Integrate a spline from a knot at i to the value x.

(((A(1/4)cs.c[4, i] * Δ + A(1/3)cs.c[3, i]) * Δ + A(1/2)cs.c[2, i]) * Δ + cs.c[1, i])Δ

Glissa.integrate — Method

Integrate a spline from x1 to x2.

Glissa.normalized_bspline — Method

An explicit representation of the normalized B-spline.

Glissa.project_to_monotonicity! — Method

Uses Hyman piecewise monotonicity.

Glissa.single_polynomial — Method

Construct a polynomial that interpolates a function, using polynomials with zeroes. This should match the constructor above. It comes from deBoor equation 4.2.