Reference

Glissa.DividedDifference

Glissa.NewtonForm

Glissa.NewtonForm

Glissa.PiecewisePolynomial

Glissa.RepeatedIndex

Base.:*

Base.eltype

Base.length

Glissa.bspline_by_matrix!

Glissa.bsplineq_recursive

Glissa.bsplineq_recursive

Glissa.cubic_spline

Glissa.cubic_spline_coefficients!

Glissa.deboor_swartz_criterion

Glissa.degree

Glissa.derivative

Glissa.divdiff

Glissa.divdiff_repeats

Glissa.divided_difference

Glissa.divided_difference_explicit

Glissa.dprod

Glissa.evaluate!

Glissa.evaluate_bspline

Glissa.generate_normalized_bsplines!

Glissa.global_derivatives!

Glissa.hyman_coefficients!

Glissa.hyman_criterion

Glissa.integral_in_interval

Glissa.integrate

Glissa.normalized_bspline

Glissa.order

Glissa.piecewise_representation!

Glissa.polyspline_constraints!

Glissa.project_to_monotonicity!

Glissa.reduce_axis

Glissa.single_polynomial

Glissa.splinen_divided

Glissa.splineq_divided

Glissa.truncated

Glissa.truncated

Glissa.vandermonde_determinant_repeats

Glissa.DividedDifference — Type

A recursive implementation of divided differences.

Glissa.NewtonForm — Type

The Newton form for an interpolating polynomial represents $f(x)$ at points $\tau$ by a sum.

$\sum_{i=0}^n f[x_0,\ldots,x_i]\prod_{j=0}^{i-1}(x-x_j)$

Here, the a are the divided differences and $\tau$ are the x[j].

Glissa.NewtonForm — Method

Construct a NewtonForm. From Conte and deBoor 1980, Eqn. 4.8.

Glissa.PiecewisePolynomial — Type

This is a piecewise polynomial. It has N+1 points on the abcissa, $\tau$. It has N cubic interpolants in c, which is size 4 x N for a cubic spline. At each interval, the polynomial is in Horner Form,

$P_i(x) = c_{1,i} + c_{2,i}(x-x_i) + c_{3,i}(x-x_i)^2 + c_{4,i}(x-x_i)^3$

This spline has one type for the abcissa and one for the ordinate coefficients. It should be the case that one(T) * one(X) is of type T.

This can be a spline of another order. It's just a matter of the dimensions of the coefficient array, c. You would evaluate this with evalpoly(x-τ[i], c[1:end, i]), where i is the interval.

Glissa.RepeatedIndex — Type

RepeatedIndex(multiple::AbstractVector)

Uses a multiplicity vector to present a vector as having repeated elements. If a vector is length 3, and multiple is [2, 1, 3], then this index will return the values[1, 1, 2, 3, 3, 3]`.

Base.:* — Method

Multiply two splines of order m1 and m2 to get a spline of order m1+m2. This only works if the two splines have the same abcissa.

Base.eltype — Method

Type of the polynomial constants.

Base.length — Method

Number of vertices in axis, which is 1 + number of intervals

Glissa.bspline_by_matrix! — Method

Calculate coefficients of a b-spline of order m, degree m-1 by sending an axis of length m+1 and a coefficient matrix of size mxm, into which are written columns of coefficients for the polynomial of each interval. If you want a b-spline on a subset of an axis, use an array view for inputs. This is the least efficient way to calculate these coefficients, but it makes explicit the continuity conditions that lead to a solution. i is the index of the leftmost point, counting points by their multiplicity, so a point with multiplicity of 3 is three points.

Glissa.bsplineq_recursive — Method

bsplineq(axis, order, index, x)

This is the $Q_i^m(x)$ B-spline, which is normalized to $1/m$. Here is is defined by recursion. Needs length(axis) ≤ index + order. The order is 1 or greater.

Glissa.bsplineq_recursive — Method

bsplineq(axis, order, index, x, nth_derivative)

This is the $Q_i^m(x)$ B-spline, which is normalized to $1/m$. Here is is defined by recursion.

Glissa.cubic_spline — Method

Creates a cubic spline on abcissa τ with values f and endpoint slopes equal to fp[1] and fp[2].

Glissa.cubic_spline_coefficients! — Method

Given an abcissa, τ, of length N+1 values f of length N+1, and derivatives s of length N+1, this computes the 4xN matrix c of cubic coefficients. Conte and deBoor Eq. 4.55

Glissa.deboor_swartz_criterion — Method

"Accurate Monotonicity Preserving Cubic Interpolation" by James M Hyman. 1983.

Glissa.degree — Method

Polynomial degree, so 2 for $x^2$

Glissa.derivative — Method

Given a spline, create a new spline that is its derivative.

Glissa.divdiff — Method

Definition 2.49 of divided differences from Schumaker. Eqn 2.86.

axis = 0:5 f(x) = 2.3 + 0.5*x^2 - x^3 divdiff(axis, f)

Glissa.divdiff_repeats — Method

Divided differences that allow for derivatives. Schumaker page 46-7, eqn 2.89.

Glissa.divided_difference — Method

Calculate divided differences of a function f at points τ, using recursion.

\[[\tau_i, \tau_{i+1},\ldots,\tau_j]f\]

This calculates the divided difference from τi to τj. The dd is a memoization. If any knots are equal, define f(x,n) where n is how many times to take the derivative.

Glissa.divided_difference_explicit — Method

This explicit divided difference applies only when the axis is distinct.

Glissa.dprod — Method

$a(a+1)(a+2)\cdots(b-1)b$

Glissa.evaluate! — Method

Evaluate a piecewise polynomial at a series of x values and write into y.

Glissa.evaluate_bspline — Method

evaluate_bspline(c::AbstractArray{T}, y::AbstractArray, m, x::T) where {T}

Evaluate B-spline with coefficients c on axis y of order m at x. This is Algorithm 5.8 from Schumaker.

Glissa.generate_normalized_bsplines! — Method

generate_normalized_bsplines!(N::AbstractVector{T}, y::AbstractVector, l, m, x::T) where {T}

Given x in $[y_l, y_l+1)$ to generate $N^m_{l+1-m}(x)$ to $N_l^m(x).$ y is the axis. m is the order of the b-spline. The l is found by a search of the axis, and the axis can have repeated elements. This algorithm applies to axis intervals where $y_{l+1} > y_l$, not equal to it. That tells us how to number the expansion coefficients, c. It's Algorithm 5.5 from Schumaker's book.

Glissa.global_derivatives! — Method

τ are the abcissa from 1 to N+1. f are the known values of length N+1. fp_bounds are derivatives at the first and last endpoint. fp is the output, of length N + 1. This is from Conte and deBoor Sec. 4.8, Eq. 4.60. The equation is this, with $i$ running from 2 to N.

$(\Delta x_i) s_{i-1}+2(\Delta x_{i-1}+\Delta x_i)s_i +(\Delta x_{i-1}s_{i+1}) = 3(f[x_{i-1},x_i]\Delta x_i+f[x_i,x_{i+1}]\Delta x_{i-1})$

Here, $\Delta x_i = x_{i+1} -x{i}$ and

$f[x_i,x_{i+1}] = \frac{f[x_{i+1}] - f[x_i]}{x_{i+1}-x_i}$

Glissa.hyman_coefficients! — Method

This is how splinefun sets the coefficients. Seems to have the same result as the code above.

Glissa.hyman_criterion — Method

This comes from Hyman's paper but is modified by the R implementation of splinefun, which Simon Wood wrote. I didn't understand from the paper the exact structure of decisions to take in the function.

Glissa.integral_in_interval — Method

Integrate a spline from a knot at i to the value x.

(((A(1/4)cs.c[4, i] * Δ + A(1/3)cs.c[3, i]) * Δ + A(1/2)cs.c[2, i]) * Δ + cs.c[1, i])Δ

Glissa.integrate — Method

Integrate a spline from x1 to x2.

Glissa.normalized_bspline — Method

An explicit representation of the normalized B-spline.

Glissa.order — Method

The order is one plus the degree of the polynomial.

Glissa.piecewise_representation! — Method

Algorithm 5.15: Construct a piecewise polynomial representation of a polynomial spline that is represented by spline coefficients. This converts out of B-splines to the polynomial representation. If there are more than about 2 evaluations of each spline interval, this can be more efficient. Each spline $i$ is represented by a polynomial of the form

$s_i(x)= \sum_{j=1}^m w_{ji} (x - \tau_i)^{j-1}$

where $w_{ij}$ are polynomial constants for the ith B-spline.

Initialize w = zeros(T, m, n) where n is the number of B-spline coefficients.

Glissa.polyspline_constraints! — Method

This reads a list of multiplicities and converts that into a matrix that represents the constraints indicated by those multiplicites. The resulting matrix defines the space of the given polynomial spline on this axis.

Glissa.project_to_monotonicity! — Method

Uses Hyman piecewise monotonicity.

Glissa.reduce_axis — Method

This extracts from a multiplicity vector the support for the ith B-spline. There is an axis where length(axis)=length(multiplicity). This axis will, in general, support multiple B-splines. This function returns a sub-range of that axis and a new multiplicity vector, the same length as that sub-range. It will be the exact support for the ith B-spline of order order.

Glissa.single_polynomial — Method

Given an axis τ and values f on that axis, return a polynomial y such that

$y(τ[i]) = f[i]$

Construct a polynomial that interpolates a function, using polynomials with zeroes. This should match the constructor above. It comes from deBoor equation 4.2.

Glissa.splinen_divided — Method

An N version of the B-spline on an axis y, order m, index i, at x. This version is defined using divided differences.

Glissa.splineq_divided — Method

A Q version of the B-spline on an axis y, order m, index i, at x. This version is defined using divided differences.

Glissa.truncated — Method

A truncated function with a derivative with respect to c.

Glissa.truncated — Method

This is the truncated function that defines the Q B-spline.

\[(x - c)_{+}^k\]

It's only defined when $x > c$.

Glissa.vandermonde_determinant_repeats — Method

The equation 2.67 in Schumaker is hard to understand. It reads, exactly,

\[V(t_1,\ldots,t_m)=\prod_{1\le i<j\le d}(\tau_j-\tau_i)^{l_jl_i} \prod_{i=1}^d\prod_{\nu=1}^{l_i-1}\nu!\]

The LHS goes to $m$ to count the total length of the axis. The RHS uses $d$ which is the number of unique values. I have a hard time believing it's a product of factorials at the end. Maybe he wanted just a single product over the factorial?

\[\prod_{i=1}^d (\l_i-1)!\]

Let's check it.